task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -14,7 +14,7 @@ int main(void)
 	long int nt = 3;
 
 	printf(" %ld, %ld, ", t1, t2);
-	int x;
+	long int x;
 
 	for (x = 3; x < 50; x++)
